feat(cookies): add build-tag-selectable cookie provider (kooky default, HackBrowserData opt-in)
Split internal/cookies into a provider-agnostic core plus two readers chosen at
build time, so the browser cookie backend can be swapped without touching the
selection logic.

- cookies.go keeps the pure, unit-tested selection pipeline (rawCookie,
  groupCandidates, selectSession, chooseSession, annotateReadError).
- provider_kooky.go (//go:build !hbd) is the default: browserutils/kooky. A stock
  `go build` links kooky and nothing from HackBrowserData.
- provider_hbd.go (//go:build hbd) reads via moond4rk/hackbrowserdata, with
  per-OS retriever wiring (retrievers_hbd_*.go) deliberately kept to the benign
  native path: no macOS login-password prompt, no Windows ABE reflective
  injection, standard Linux keyring.
- browserReadHints and the provider tests move to the matching tagged files.

Both tag variants build (darwin/linux/windows), vet, and test clean. go.mod lists
both deps; only the tagged one is linked per build.
